% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centervar_plot-tbl.R
\name{centervar_tbl}
\alias{centervar_tbl}
\title{Table stratified by center}
\usage{
centervar_tbl(.db, what)
}
\value{
a data frame with levels' counts for the selected variable,
each center.
}
\description{
Table stratified by center
}
