% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox_plot-tbl.R
\name{checkbox_plot}
\alias{checkbox_plot}
\title{Checkbos plot}
\usage{
checkbox_plot(.db, dict)
}
\value{
a ggplot
}
\description{
A plot checkboxed characteristics reported as proportionally filled
bars.
}
