% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox_plot-tbl.R
\name{checkbox_tbl}
\alias{checkbox_tbl}
\title{Checkbox table}
\usage{
checkbox_tbl(.db, dict)
}
\value{
a dataframe with sum and prop for each checkbox
}
\description{
A long table with sum and proportions stratifiable by gender and age
class
}
