% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-descriptivesReport.R
\name{descriptives_ageclassPlot}
\alias{descriptives_ageclassPlot}
\title{descriptives_etniaPlot(full_records)}
\usage{
descriptives_ageclassPlot(.db)
}
\description{
descriptives_etniaPlot(full_records)
}
