% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-descriptivesReport.R
\name{descriptives_etniaPlot}
\alias{descriptives_etniaPlot}
\title{descriptives_etniaPlot(full_records)}
\usage{
descriptives_etniaPlot(.db)
}
\description{
descriptives_etniaPlot(full_records)
}
