% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-descriptivesReport.R
\name{descriptives_genderPlot}
\alias{descriptives_genderPlot}
\title{descriptives_genderPlot(full_records)}
\usage{
descriptives_genderPlot(.db)
}
\description{
descriptives_genderPlot(full_records)
}
