% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-originReport.R
\name{descriptives_originPlot}
\alias{descriptives_originPlot}
\title{descriptives_originPlot(full_records)}
\usage{
descriptives_originPlot(.db)
}
\description{
descriptives_originPlot(full_records)
}
