% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-originReport.R
\name{descriptives_tipologiaPlot}
\alias{descriptives_tipologiaPlot}
\title{descriptives_tipologiaPlot(full_records)}
\usage{
descriptives_tipologiaPlot(.db)
}
\description{
descriptives_tipologiaPlot(full_records)
}
