% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-admissionReport.R
\name{funs-admissionReport}
\alias{funs-admissionReport}
\alias{admission_Plot}
\alias{admission_dataToUse}
\title{picu admissions module}
\usage{
admission_Plot(
  .db,
  type = c("centervar", "checkbox"),
  what = c("mal_cronica", "popc", "insuff_organo7"),
  dict = NULL
)

admission_dataToUse(.db, .which = c("Completed only", "Overall"))
}
\arguments{
\item{.db}{(data frame) data}

\item{what}{(chr, default "gender") variable to describe. One of
"gender", or "etnicity"}
}
\description{
picu admissions module
}
\section{Functions}{
\itemize{
\item \code{admission_Plot()}: plot

\item \code{admission_dataToUse()}: data

}}
\examples{

admission_Plot(full_records) # default what = "mal_cronica"
admission_Plot(full_records, "mal_cronica") # same as before
admission_Plot(full_records, "etnia")

admission_dataToUse(full_records, "Completed only")
admission_dataToUse(full_records, "overall")
}
