% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-admissionReport1.R
\docType{data}
\name{funs-admissionReport1}
\alias{funs-admissionReport1}
\alias{comorb}
\title{picu admissions module}
\format{
An object of class \code{character} of length 13.
}
\usage{
comorb
}
\arguments{
\item{.db}{(data frame) data
"gender", or "etnicity"}
}
\description{
picu admissions module
}
\section{Functions}{
\itemize{
\item \code{comorb}: plot

}}
\examples{

admission_Plot1(full_records, "mal_cronica") # same as before
admission_Plot1(full_records, "etnia")
}
\keyword{datasets}
