% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-focusReport.R
\name{funs-focusReport}
\alias{funs-focusReport}
\alias{focus_Plot}
\alias{focus_dataToUse}
\title{picu focus module}
\usage{
focus_Plot(
  .db,
  what = c("insufficienza", "procedure_rico", "vent_iniz2", "niv_it", "inf_ingresso_tip",
    "sede_inf2", "diagnosi_inf", "tipo_inf", "sepsi_2")
)

focus_dataToUse(.db, .which = c("Completed only", "Overall"))
}
\arguments{
\item{.db}{(data frame) data}

\item{what}{(chr, default "gender") variable to describe. One of
"gender", or "etnicity"}
}
\description{
picu focus module
}
\section{Functions}{
\itemize{
\item \code{focus_Plot()}: plot

\item \code{focus_dataToUse()}: data

}}
\examples{

focus_Plot(full_records) # default what = "mal_cronica"
focus_Plot(full_records, "mal_cronica") # same as before
focus_Plot(full_records, "etnia")

focus_dataToUse(full_records, "Completed only")
focus_dataToUse(full_records, "overall")
}
