% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-losReport.R
\name{funs-losReport}
\alias{funs-losReport}
\alias{los_Plot}
\alias{pimlos_plot}
\alias{smrlos_plot}
\alias{smrlosvol_plot}
\alias{los_dataToUse}
\title{los module}
\usage{
los_Plot(
  .db,
  what = c("diagnosi", "diagnosi_2", "popc_dimissione", "popc_delta", "esito_tip",
    "mod_decesso", "deceduto", "prelievo_organi", "destinazione", "durata_degenza",
    "smr")
)

pimlos_plot(.db)

smrlos_plot(.db)

smrlosvol_plot(.db)

los_dataToUse(.db, .which = c("Completed only", "Overall"))
}
\arguments{
\item{.db}{(data frame) data}

\item{what}{(chr, default "gender") variable to describe. One of
"gender", or "etnicity"}
}
\description{
los module
}
\section{Functions}{
\itemize{
\item \code{los_Plot()}: plot

\item \code{pimlos_plot()}: pim-time plot

\item \code{smrlos_plot()}: smr-time plot

\item \code{smrlosvol_plot()}: smr-vol plot

\item \code{los_dataToUse()}: data

}}
\examples{

\dontrun{
  los_Plot(full_records) # default what = "mal_cronica"
  los_Plot(full_records, "mal_cronica") # same as before
  los_Plot(full_records, "etnia")
}

\dontrun{
  los_dataToUse(full_records, "Completed only")
  los_dataToUse(full_records, "overall")
}
}
