% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-originReport.R
\name{funs-originReport}
\alias{funs-originReport}
\alias{origin_Plot}
\alias{origin_dataToUse}
\title{Origin module}
\usage{
origin_Plot(
  .db,
  what = c("ricovero_progr", "redcap_repeat_instance"),
  by_ageclass = FALSE,
  by_gender = FALSE
)

origin_dataToUse(.db, .which = c("Completed only", "Overall"))
}
\arguments{
\item{.db}{(data frame) data}

\item{what}{(chr, default "gender") variable to describe. One of
"gender", or "etnicity"}
}
\description{
Origin module
}
\section{Functions}{
\itemize{
\item \code{origin_Plot()}: plot

\item \code{origin_dataToUse()}: data

}}
\examples{

origin_Plot(full_records) # default what = "gender"
origin_Plot(full_records, "gender") # same as before
origin_Plot(full_records, "etnia")

origin_dataToUse(full_records, "Completed only")
origin_dataToUse(full_records, "overall")
}
