% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-admissionReport1.R
\name{module-admissionReport1}
\alias{module-admissionReport1}
\alias{admissionReportUI1}
\title{Admission module}
\usage{
admissionReportUI1(id)
}
\arguments{
\item{id}{name for the specific instance of the module.}

\item{data}{database to use}

\item{completed}{(chr) "Completed" or "Not-completed"}

\item{type}{(chr) "Total" or "Proportion"}
}
\description{
General description
}
\section{Functions}{
\itemize{
\item \code{admissionReportUI1()}: user interface

}}
