% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-descriptivesReport.R,
%   R/module-originReport.R
\name{module-descriptivesReport}
\alias{module-descriptivesReport}
\alias{descriptivesReportUI_gender}
\alias{descriptivesReport_gender}
\alias{descriptivesReportStatic_gender}
\alias{descriptivesReportUI_etnia}
\alias{descriptivesReport_etnia}
\alias{descriptivesReportStatic_etnia}
\alias{descriptivesReportUI_ageclass}
\alias{descriptivesReport_ageclass}
\alias{descriptivesReportStatic_ageclass}
\alias{OriginReport_origin}
\alias{OriginReportStatic_origin}
\alias{OriginReport_tipologia}
\alias{OriginReportStatic_tipologia}
\title{Descriptives module}
\usage{
descriptivesReportUI_gender(id)

descriptivesReport_gender(id, data)

descriptivesReportStatic_gender(data, completed)

descriptivesReportUI_etnia(id)

descriptivesReport_etnia(id, data)

descriptivesReportStatic_etnia(data, completed)

descriptivesReportUI_ageclass(id)

descriptivesReport_ageclass(id, data)

descriptivesReportStatic_ageclass(data, completed)

OriginReport_origin(id, data)

OriginReportStatic_origin(data, completed)

OriginReport_tipologia(id, data)

OriginReportStatic_tipologia(data, completed)
}
\arguments{
\item{id}{name for the specific instance of the module.}

\item{data}{database to use}
}
\description{
General description
}
\section{Functions}{
\itemize{
\item \code{descriptivesReportUI_gender}: user interface

\item \code{descriptivesReport_gender}: server function

\item \code{descriptivesReportStatic_gender}: static report function

\item \code{descriptivesReportUI_etnia}: user interface

\item \code{descriptivesReport_etnia}: server function

\item \code{descriptivesReportStatic_etnia}: static report function

\item \code{descriptivesReportUI_ageclass}: user interface

\item \code{descriptivesReport_ageclass}: server function

\item \code{descriptivesReportStatic_ageclass}: static report function

\item \code{OriginReport_origin}: server function

\item \code{OriginReportStatic_origin}: static report function

\item \code{OriginReport_tipologia}: server function

\item \code{OriginReportStatic_tipologia}: static report function
}}

