% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-focusReport.R
\name{module-focusReport}
\alias{module-focusReport}
\alias{focusReportUI}
\alias{focusReport}
\alias{focusReportStatic}
\title{focus module}
\usage{
focusReportUI(id)

focusReport(id, data, what)

focusReportStatic(data, completed, what)
}
\arguments{
\item{id}{name for the specific instance of the module.}

\item{data}{database to use}

\item{completed}{(chr) "Completed" or "Not-completed"}

\item{type}{(chr) "Total" or "Proportion"}
}
\description{
General description
}
\section{Functions}{
\itemize{
\item \code{focusReportUI()}: user interface

\item \code{focusReport()}: server function

\item \code{focusReportStatic()}: static report function

}}
