% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-lossmrReport.R
\name{module-lossmrReport}
\alias{module-lossmrReport}
\alias{lossmrReportUI}
\alias{lossmrReport}
\title{lossmr module}
\usage{
lossmrReportUI(id)

lossmrReport(id, data)
}
\arguments{
\item{id}{name for the specific instance of the module.}

\item{data}{database to use}
}
\description{
General description
}
\section{Functions}{
\itemize{
\item \code{lossmrReportUI()}: user interface

\item \code{lossmrReport()}: server function

}}
