% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-lossmrvolReport.R
\name{module-lossmrvolReport}
\alias{module-lossmrvolReport}
\alias{lossmrvolReportUI}
\alias{lossmrvolReport}
\title{lossmrvol module}
\usage{
lossmrvolReportUI(id)

lossmrvolReport(id, data)
}
\arguments{
\item{id}{name for the specific instance of the module.}

\item{data}{database to use}
}
\description{
General description
}
\section{Functions}{
\itemize{
\item \code{lossmrvolReportUI()}: user interface

\item \code{lossmrvolReport()}: server function

}}
