% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-originReport.R
\name{module-originReport}
\alias{module-originReport}
\alias{OriginReportUI_origin}
\alias{OriginReportUI_tipologia}
\title{Origin module}
\usage{
OriginReportUI_origin(id)

OriginReportUI_tipologia(id)
}
\arguments{
\item{id}{name for the specific instance of the module.}

\item{data}{database to use}
}
\description{
General description
}
\section{Functions}{
\itemize{
\item \code{OriginReportUI_origin}: user interface

\item \code{OriginReportUI_tipologia}: user interface
}}

