% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipnet.ubesp-package.R
\docType{package}
\name{tipnet.report-package}
\alias{tipnet.report}
\alias{tipnet.report-package}
\title{tipnet.report: TIPNet reports}
\description{
The goal of tipnet.report is to implement and collect all the code needed for the development of the TIPNet reports by UBESP.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UBESP-DCTV/tipnet.report}
  \item Report bugs at \url{https://github.com/UBESP-DCTV/tipnet.report/issues}
}

}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@gmail.com} [copyright holder]

}
\keyword{internal}
