% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-admissionReport1.R
\name{funs-admissionReport1}
\alias{funs-admissionReport1}
\alias{admission_Plot1}
\title{picu admissions module}
\usage{
admission_Plot1(.db, dict)
}
\arguments{
\item{.db}{(data frame) data
"gender", or "etnicity"}
}
\description{
picu admissions module
}
\section{Functions}{
\itemize{
\item \code{admission_Plot1()}: plot

}}
\examples{

admission_Plot1(full_records, "mal_cronica") # same as before
admission_Plot1(full_records, "etnia")
}
