% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-originReport.R
\name{funs-OriginReport}
\alias{funs-OriginReport}
\alias{descriptives_dataToUse}
\title{Origin module}
\usage{
descriptives_dataToUse(.db, .which = c("Completed only", "Overall"))
}
\arguments{
\item{.db}{(data frame) data}
}
\description{
Origin module
}
\section{Functions}{
\itemize{
\item \code{descriptives_dataToUse}: plot
}}

\examples{

#' @describeIn funs-OriginReport data

descriptives_dataToUse(full_records, "Completed only")
descriptives_dataToUse(full_records, "overall")
}
