% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-admissionReport.R,
%   R/module-admissionReport1.R
\name{module-admissionReport}
\alias{module-admissionReport}
\alias{admissionReportUI}
\alias{admissionReport}
\alias{admissionReportStatic}
\alias{admissionReport1}
\alias{admissionReportStatic1}
\title{Admission module}
\usage{
admissionReportUI(id)

admissionReport(id, data, what)

admissionReportStatic(data, completed, what)

admissionReport1(id, data)

admissionReportStatic1(data, completed)
}
\arguments{
\item{id}{name for the specific instance of the module.}

\item{data}{database to use}

\item{completed}{(chr) "Completed" or "Not-completed"}

\item{type}{(chr) "Total" or "Proportion"}
}
\description{
General description
}
\section{Functions}{
\itemize{
\item \code{admissionReportUI()}: user interface

\item \code{admissionReport()}: server function

\item \code{admissionReportStatic()}: static report function

\item \code{admissionReport1()}: server function

\item \code{admissionReportStatic1()}: static report function

}}
