% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-descriptivesReport.R
\name{module-descriptivesReport}
\alias{module-descriptivesReport}
\alias{descriptivesReportUI}
\alias{descriptivesReport}
\alias{descriptivesReportStatic}
\title{Descriptives module}
\usage{
descriptivesReportUI(id, what = NULL)

descriptivesReport(id, data, what)

descriptivesReportStatic(data, completed, what)
}
\arguments{
\item{id}{name for the specific instance of the module.}

\item{data}{database to use}
}
\description{
General description
}
\section{Functions}{
\itemize{
\item \code{descriptivesReportUI()}: user interface

\item \code{descriptivesReport()}: server function

\item \code{descriptivesReportStatic()}: static report function

}}
