% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-lospimReport.R
\name{module-lospimReport}
\alias{module-lospimReport}
\alias{lospimReportUI}
\alias{lospimReport}
\alias{lospimReportStatic}
\title{lospim module}
\usage{
lospimReportUI(id)

lospimReport(id, data, what)

lospimReportStatic(data, completed, what)
}
\arguments{
\item{id}{name for the specific instance of the module.}

\item{data}{database to use}

\item{completed}{(chr) "Completed" or "Not-completed"}

\item{type}{(chr) "Total" or "Proportion"}
}
\description{
General description
}
\section{Functions}{
\itemize{
\item \code{lospimReportUI()}: user interface

\item \code{lospimReport()}: server function

\item \code{lospimReportStatic()}: static report function

}}
