% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-originReport.R
\name{module-originReport}
\alias{module-originReport}
\alias{originReportUI}
\alias{originReport}
\alias{originReportStatic}
\title{Origin module}
\usage{
originReportUI(id)

originReport(id, data, what)

originReportStatic(data, completed, what)
}
\arguments{
\item{id}{name for the specific instance of the module.}

\item{data}{database to use}

\item{completed}{(chr) "Completed" or "Not-completed"}

\item{type}{(chr) "Total" or "Proportion"}
}
\description{
General description
}
\section{Functions}{
\itemize{
\item \code{originReportUI}: user interface

\item \code{originReport}: server function

\item \code{originReportStatic}: static report function
}}

